requirements.txt
accelerate>=1.12.0
av>=16.0.1
einops>=0.8.1
fastapi>=0.115.0
hydra-colorlog>=1.2.0
hydra-core>=1.3.2
pandas>=2.3.3
physical_ai_av>=0.1.0
pillow>=12.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
torchvision>=0.23.0
transformers==4.57.1
uvicorn[standard]>=0.30.0
